Add files via upload
This was a team project completed with 3 others in 1st year. The brief was to design a website that could be accessed from a kiosk inside the Free Derry Museum in Derry. The website/kiosk was intended to display historical artifacts for which there was no space on the museum floor, to date exclusively posters. 2 folders of scanned posters were given with the brief.